+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 
 ### Experimental
 - [CSE] Added check if the operation execution time (oet) is after rqet or rset.
+- [CSE] Added exeperimental support for the [credential] ManagementObject specialization.
 
 ### Changed